commandHauls: extract networking to module
